src/VumbaSoft.AdventureWorks.Controllers/Production/ProductDescriptions/ProductDescriptionsController.cs
using Microsoft.AspNetCore.Mvc;
using VumbaSoft.AdventureWorks.Objects;
using VumbaSoft.AdventureWorks.Services;
using VumbaSoft.AdventureWorks.Validators;
using System;

namespace VumbaSoft.AdventureWorks.Controllers.Production
{
    [Area("Production")]
    public class ProductDescriptionsController : ValidatedController<IProductDescriptionValidator, IProductDescriptionService>
    {
        public ProductDescriptionsController(IProductDescriptionValidator validator, IProductDescriptionService service)
            : base(validator, service)
        {
        }

        [HttpGet]
        public ViewResult Index()
        {
            return View(Service.GetViews());
        }

        [HttpGet]
        public ViewResult Create()
        {
            return View();
        }

        [HttpPost]
        public ActionResult Create(ProductDescriptionView description)
        {
            if (!Validator.CanCreate(description))
                return View(description);

            Service.Create(description);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Details(Int32 id)
        {
            return NotEmptyView(Service.Get<ProductDescriptionView>(id));
        }

        [HttpGet]
        public ActionResult Edit(Int32 id)
        {
            return NotEmptyView(Service.Get<ProductDescriptionView>(id));
        }

        [HttpPost]
        public ActionResult Edit(ProductDescriptionView description)
        {
            if (!Validator.CanEdit(description))
                return View(description);

            Service.Edit(description);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Delete(Int32 id)
        {
            return NotEmptyView(Service.Get<ProductDescriptionView>(id));
        }

        [HttpPost]
        [ActionName("Delete")]
        public RedirectToActionResult DeleteConfirmed(Int32 id)
        {
            Service.Delete(id);

            return RedirectToAction("Index");
        }
    }
}
